Remove the githubfilelogo.svg since it is not actually the github logo. (#53)
Infact, this is more like a full HTML file than an a simple SVG. Not only
that, this file tripped CredScan since it included a CSRF token. CSRF tokens
are meant to be sessions-specific and should not have been saved in a static
file. Loading this svg file just results in the text
"pxt-adafruit/githubfilelogo.svg at master · Microsoft/pxt-adafruit"
being rendered in the browser, which has nothing to do with this repo. All
in all, there does not appear to be a reason to keep this file around.